azurestorage_blueprints/requirements.txt
azure-storage-blob==12.3.2
